Extract portfolio workspace contract family
